refactor: update keySuffixOptions filename
