+++ .github/whitelist-capital-letters.txt
@@ -40,7 +40,6 @@
 ./libs/common/src/misc/serviceUtils.spec.ts
 ./libs/common/src/enums/linkedIdType.ts
 ./libs/common/src/enums/notificationType.ts
-./libs/common/src/enums/keySuffixOptions.ts
 ./libs/common/src/enums/productType.ts
 ./libs/common/src/enums/uriMatchType.ts
 ./libs/common/src/enums/secureNoteType.ts
