+++ apps/cli/src/commands/serve.command.ts
@@ -5,7 +5,7 @@ import * as koa from "koa";
 import * as koaBodyParser from "koa-bodyparser";
 import * as koaJson from "koa-json";
 
-import { KeySuffixOptions } from "@bitwarden/common/enums/keySuffixOptions";
+import { KeySuffixOptions } from "@bitwarden/common/enums/key-suffix-options";
 import { Utils } from "@bitwarden/common/misc/utils";
 
 import { ConfirmCommand } from "../admin-console/commands/confirm.command";
