+++ libs/common/src/enums/key-suffix-options.ts
